refactor: projectStraight domain 패키지 변경
