+++ src/main/java/baekgwa/suhoserver/domain/project/service/ProjectReadService.java
@@ -15,8 +15,8 @@ import baekgwa.suhoserver.model.project.branch.entity.ProjectBranchEntity;
 import baekgwa.suhoserver.model.project.branch.repository.ProjectBranchRepository;
 import baekgwa.suhoserver.model.project.project.entity.ProjectEntity;
 import baekgwa.suhoserver.model.project.project.repository.ProjectRepository;
-import baekgwa.suhoserver.model.project.straight.entity.ProjectStraightEntity;
-import baekgwa.suhoserver.model.project.straight.repository.ProjectStraightRepository;
+import baekgwa.suhoserver.model.project.straight.straight.entity.ProjectStraightEntity;
+import baekgwa.suhoserver.model.project.straight.straight.repository.ProjectStraightRepository;
 import lombok.RequiredArgsConstructor;
 
 /**
