+++ src/main/java/baekgwa/suhoserver/model/project/straight/straight/repository/ProjectStraightRepository.java
@@ -1,4 +1,4 @@
-package baekgwa.suhoserver.model.project.straight.repository;
+package baekgwa.suhoserver.model.project.straight.straight.repository;
 
 import java.util.List;
 
@@ -7,10 +7,10 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
 import baekgwa.suhoserver.model.project.project.entity.ProjectEntity;
-import baekgwa.suhoserver.model.project.straight.entity.ProjectStraightEntity;
+import baekgwa.suhoserver.model.project.straight.straight.entity.ProjectStraightEntity;
 
 /**
- * PackageName : baekgwa.suhoserver.model.project.straight.repository
+ * PackageName : baekgwa.suhoserver.model.project.straight.straight.repository
  * FileName    : ProjectStraightRepository
  * Author      : Baekgwa
  * Date        : 2025-08-09
@@ -25,5 +25,6 @@ public interface ProjectStraightRepository extends JpaRepository<ProjectStraight
 	List<ProjectStraightEntity> findByProject(ProjectEntity findProject);
 
 	@Query("SELECT ps FROM ProjectStraightEntity ps JOIN FETCH ps.straightType st WHERE ps.project = :project AND ps.isLoopRail = :isLoopRail ORDER BY ps.length DESC , st.type ASC")
-	List<ProjectStraightEntity> findSortedWithType(@Param("project") ProjectEntity project, @Param("isLoopRail") Boolean isLoopRail);
+	List<ProjectStraightEntity> findSortedWithType(@Param("project") ProjectEntity project,
+		@Param("isLoopRail") Boolean isLoopRail);
 }
